requirements.txt
python-telegram-bot[webhooks]==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
apscheduler==3.10.4
pytz==2023.3
openai==1.3.8
redis==5.0.1
loguru==0.7.2
babel==2.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiofiles==23.2.1
cryptography==41.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
greenlet==3.0.3
matplotlib==3.8.2
httpx==0.25.2